protobuf: rename package to 'naisdevice' with go alias 'pb'
